Fix AWS deploy error
I noticed that all deployments would sometimes fail when a change
was first made. To fix it, it would need to be redeloyed via CodeDeploy.
This happened becuase sometimes the server would hang for too long,
awaiting input, but it would still deploy and run the code.

This fix ensures the bug doesn't happen and all input and output is
run in a background process versus stdin/stdout.